* Implement C_CLASS_CHANGE and C_LEADERBOARD packets (note: both untested) * Finish implementing C_ROOM_CREATE and C_ROOM_JOIN packets * Send appropriate error-responses for some packets * Change the SQL database to use MEDIUMINT instead of INT and reflect this in the java code (replacing longs with ints)     - TODO: Upload create_sql queries

+++ trunk/src/server/KatanaClient.java
@@ -13,7 +13,7 @@ import shared.KatanaPacket;
 
 public class KatanaClient implements Runnable
 {
-    private long id;
+    private int id;
     private Socket client;
     Thread thread;
     
@@ -27,7 +27,7 @@ public class KatanaClient implements Runnable
         thread.start();
     }
     
-    public void setId(long id) { this.id = id; }
+    public void setId(int id) { this.id = id; }
     public long getId() { return id; }
     
     public void remove()
